% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_api_key}
\alias{check_api_key}
\title{check_api_key}
\usage{
check_api_key()
}
\value{
Error message if api_key is absent
}
\description{
Generic function to check that the user has supplied an API key for twfy
}
