% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_texts_fun}
\alias{clean_texts_fun}
\title{clean_texts_fun}
\usage{
clean_texts_fun(string)
}
\value{
Clean speech string
}
\description{
Generic function to clean up a speech string (remove html elements, etc)
}
