% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dates}
\alias{convert_dates}
\title{convert_dates}
\usage{
convert_dates(x)
}
\arguments{
\item{x}{Date string to convert}
}
\value{
R date object
}
\description{
Generic function to convert TWFY dates to R dates
}

