% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dates_r_to_twfy}
\alias{convert_dates_r_to_twfy}
\title{convert_dates_r_to_twfy}
\usage{
convert_dates_r_to_twfy(x)
}
\arguments{
\item{x}{R date object to convert}
}
\value{
Character object
}
\description{
Generic function to convert R dates to twfy dates
}
