% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_args_global}
\alias{func_args_global}
\title{func_args_global}
\usage{
func_args_global(func)
}
\value{
A bunch of R objects
}
\description{
Function to take a function and set the default arguments of the function as objects in the global environment for testing. Don't use this unless you wrote this horrible code.
}

