% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_generic.R
\name{get_generic}
\alias{get_generic}
\title{get_generic}
\usage{
get_generic(call = "getMPs", search_date = NA, search_name = NA,
  search_postcode = NA, search_party = NA, search_string = NA,
  search_id = NA, search_constituency = NA, search_always_return = NA,
  search_fields = NA, search_gid = NA, search_order = NA,
  search_page = NA, search_num = NA, search_type = NA,
  search_person = NA, return_url = F)
}
\arguments{
\item{call}{Which api command do you want to call?}

\item{search_date}{Date paramenter}

\item{search_name}{Name paramenter}

\item{search_postcode}{Postcode paramenter}

\item{search_party}{Party paramenter}

\item{search_string}{String paramenter}

\item{search_id}{ID paramenter}

\item{search_constituency}{Constituency paramenter}

\item{search_always_return}{Always return paramenter}

\item{search_fields}{Fields paramenter}

\item{search_gid}{GID paramenter}

\item{search_order}{order paramenter}

\item{search_page}{page paramenter}

\item{search_num}{num paramenter}

\item{search_type}{type parameter}

\item{search_person}{person parameter}
}
\value{
XML data returned from TheyWorkForYou.
}
\description{
Generic function to build a TheyWorkForYou url and call in the XML data. For internal use in twfyR only. I wouldn't bother using this directly.
}

