% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundary.R
\name{getBoundary}
\alias{getBoundary}
\title{getBoundary}
\usage{
getBoundary(name = NA, plot_boundary = F)
}
\arguments{
\item{name}{Name of the constituency.}
}
\value{
A SpatialPolygonsDataFrame from rgdal::readOGR
}
\description{
Returns KML file for a UK Parliament constituency.
}
\details{
Returns the bounding polygon of the constituency, in KML format (see mapit.mysociety.org for other formats, past constituency boundaries, and so on).
}
\examples{
getBoundary("Hornsey and Wood Green", plot_boundary = T)

}

