% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstituencies.R
\name{getConstituencies}
\alias{getConstituencies}
\title{getConstituencies}
\usage{
getConstituencies(date = NA, string = NA)
}
\arguments{
\item{date}{Fetch the list of constituencies as on this date.}

\item{string}{Fetch the list of constituencies that match this search string.}
}
\value{
character vector of constituencies
}
\description{
Fetch a list of UK Parliament constituencies. Note: only one argument can be given at present.
}
\examples{
getConstituencies()

}
