% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstituency.R
\name{getConstituency}
\alias{getConstituency}
\title{getConstituency}
\usage{
getConstituency(postcode = NA, name = NA)
}
\arguments{
\item{postcode}{Fetch the constituency with associated information for a given postcode.}

\item{name}{Fetch the data associated to the constituency with this name.}
}
\value{
data.frame of constituency info
}
\description{
Fetch a UK Parliament constituency.
}
\examples{
getConstituency()

}

