% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDebates.R
\name{getDebates}
\alias{getDebates}
\title{getDebates}
\usage{
getDebates(type = NA, date = NA, search = NA, person = NA, gid = NA,
  order = NA, page = NA, num = 1000, complete_call = F)
}
\arguments{
\item{type}{One of "commons", "westminsterhall", "lords", "scotland", or "northernireland". (required)}

\item{date}{Fetch the debates for this date.}

\item{search}{Fetch the debates that contain this term.}

\item{person}{Fetch the debates by a particular person ID.}

\item{gid}{Fetch the speech or debate that matches this GID.}

\item{order}{(optional, when using search or person) "d" for date ordering, "r" for relevance ordering.}

\item{page}{(optional, when using search or person) Page of results to return.}

\item{num}{(optional, when using search or person) Number of results to return (maximum and default is 1000).}

\item{complete_call}{The twfy API will only seemingly return 1000 results per call. Setting this parameter to TRUE will iterate over the page parameter until all results are collected. Warning: this is a quick way to use up your API call allowance.}
}
\value{
data.frame of debates
}
\description{
Fetch Debates (including Oral Questions).
}
\examples{
getDebates()

}
