% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeometry.R
\name{getGeometry}
\alias{getGeometry}
\title{getGeometry}
\usage{
getGeometry(name = NA)
}
\arguments{
\item{name}{Name of the constituency.}
}
\value{
data.frame of constituency information.
}
\description{
Returns geometry information for constituencies.
}
\details{
This currently includes, for Great Britain, the latitude and longitude of the centre point of the bounding box of the constituency, its area in square metres, the bounding box itself and the number of parts in the polygon that makes up the constituency. For Northern Ireland, as we don't have any better data, it only returns an approximate (estimated by eye) latitude and longitude for the constituency's centroid.
}
\examples{
constituency_names <- getConstituencies()
getGeometry(constituency_names[1])

}

