% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHansard.R
\name{getHansard}
\alias{getHansard}
\title{getHansard}
\usage{
getHansard(search = NA, person = NA, order = NA, page = NA,
  num = 1000, complete_call = F)
}
\arguments{
\item{search}{Fetch the data that contain this term.}

\item{person}{Fetch the debates by a particular person ID.}

\item{order}{(optional, when using search or person) "d" for date ordering, "r" for relevance ordering.}

\item{page}{(optional, when using search or person) Page of results to return.}

\item{num}{(optional, when using search or person) Number of results to return (maximum and default is 1000).}

\item{complete_call}{The twfy API will only seemingly return 1000 results per call. Setting this parameter to TRUE will iterate over the page parameter until all results are collected. Warning: this is a quick way to use up your API call allowance.}
}
\value{
data.frame of Hansard entries
}
\description{
Fetch all Hansard. Note you can only supply one of 'search' or 'person' parameters at present.
}
\examples{
getHansard()

}
