% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMP.R
\name{getMP}
\alias{getMP}
\title{getMPs}
\usage{
getMP(postcode = NA, constituency = NA, id = NA, always_return = NA)
}
\arguments{
\item{postcode}{Fetch the MP for a particular postcode (either the current one, or the most recent one, depending upon the setting of the always_return variable). This will only return their current/ most recent entry in the database, look up by ID to get full history of a person. (optional)}

\item{constituency}{The name of a constituency; we will try and work it out from whatever you give us. This will only return their current/ most recent entry in the database, look up by ID to get full history of a person. (optional)}

\item{id}{If you know the person ID for the member you want (returned from getMPs or elsewhere), this will return data for that person. This will return all database entries for this person, so will include previous elections, party changes, etc. (optional)}

\item{always_return}{For the postcode and constituency options, sets whether to always try and return an MP, even if the seat is currently vacant (due to e.g. the death of an MP, or the period before an election when there are no MPs).}
}
\value{
list of MPs and offices
}
\description{
Fetch a particular MP.
}
\examples{
getMPs()

}

