% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPs.R
\name{getMPs}
\alias{getMPs}
\title{getMPs}
\usage{
getMPs(date = NA, string = NA, party = NA)
}
\arguments{
\item{date}{Fetch the list of MPs as it was on this date. (optional)}

\item{string}{Fetch the list of MPs that match this search string in their name. (optional)}

\item{party}{Fetch the list of MPs from the given party. (optional)}
}
\value{
list of MPs and offices
}
\description{
Fetch a list of MPs.
}
\examples{
getMPs()

}
