% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPsinfo.R
\name{getMPsinfo}
\alias{getMPsinfo}
\title{getMPsinfo}
\usage{
getMPsinfo(id = NA, fields = NA)
}
\arguments{
\item{id}{The person IDs (vector of integers).}

\item{fields}{Which fields you want to return as a comma separated character string (defaults to all).}
}
\value{
list of data.frames of MPs' extra information (see TheyWorkForYou for details)
}
\description{
Fetch extra information for a particular person.
}
\examples{
getMPs()

}
