% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerson.R
\name{getPerson}
\alias{getPerson}
\title{getPerson}
\usage{
getPerson(id_person = NA)
}
\arguments{
\item{id}{If you know the person ID for the member you want (returned from getMPs or elsewhere), this will return data for that person. This will return all database entries for this person, so will include previous elections, party changes, etc.}
}
\value{
list of 2 elements. The first element is a data.frame named `member` which includes information of the person's house memberships. The second is a data.frame named `office` which details the offices that member has held.
}
\description{
Fetch a particular person.
}
\examples{
getPerson(10001)

}

