% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_twfy_key.R
\name{set_twfy_key}
\alias{set_twfy_key}
\title{set_twfy_key}
\usage{
set_twfy_key(key, testing = T)
}
\arguments{
\item{key}{users TheyWorkForYou API key}
}
\value{
None
}
\description{
Sets the value of the TheyWorkForYou API key to be used with all other twfyR functions.
}
\examples{
set_twfy_key()

}

