% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_child_search}
\alias{xml_child_search}
\title{xml_child_search}
\usage{
xml_child_search(children, string = "url")
}
\arguments{
\item{children}{children of XML object}

\item{string}{xPath expression to find}
}
\value{
character vector
}
\description{
Generic function to find elements in XML children
}

