% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_text_search}
\alias{xml_text_search}
\title{xml_text_search}
\usage{
xml_text_search(string = "member_id", x)
}
\arguments{
\item{string}{xPath expression to find}

\item{x}{XML object}
}
\value{
character vector
}
\description{
Generic function to find elements in XML
}

